pyproject.toml
[project]
name = "poly-trade-scan"
version = "0.1.0"
description = "CLI for downloading historical and listening to real-time Polymarket trades on Polygon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp",
    "eth-abi",
    "python-dotenv",
    "requests",
    "typer",
    "websockets",
]

[project.scripts]
poly = "src.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
